requirements.txt
flask
PyPDF2
pinecone
firebase_admin
python-dotenv
google-generativeai
langchain
langchain-google-genai
profanity-check

-e .